+++ app/src/visionglass/cpp/DeviceDelegateVisionGlass.h
@@ -51,8 +51,8 @@ public:
   void Resume();
   void TouchEvent(const bool aDown, const float aX, const float aY);
   void ControllerButtonPressed(const bool aDown);
-  void setHead(const float aX, const float aY, const float aZ, const float aW);
-  void setControllerOrientation(const float aX, const float aY, const float aZ, const float aW);
+  void setHead(const double aX, const double aY, const double aZ, const double aW);
+  void setControllerOrientation(const double aX, const double aY, const double aZ, const double aW);
   void CalibrateController();
 protected:
   struct State;
